--- modulo3/projeto-whatslab/src/components/SectionMensagens/StyleSectionMensagens.js
@@ -1,55 +1,0 @@
-import styled from 'styled-components'
-
-export const ContainerMensagem = styled.section`
-    background-color: orange;
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    border-radius: 10px;
-    border-color: blue;
-    width: auto;
-    height: auto;
-    padding: 5px 5px;
-    margin-left: 0;
-    margin-right: auto;
-
-    
-`
-
-export const ContainerListaMensagens = styled.ul`
-
-`
-
-export const ListaMensagem = styled.li`
-    list-style: none;
-`
-
-export const TextoSemMensagem = styled.p`
-    
-    /* margin: auto auto 0 auto; */
-`
-export const Button = styled.button`
-    cursor: pointer;
-    border: none;
-    background-color: transparent;
-`
-export const Nome = styled.h1`
-    display: flex;
-
-`
-export const TextoComMensagem = styled.p`
-
-`
-
-
-// export const button = styled.button`
-//     float: right;
-//     width: 2vw;
-//     height: 2vh;
-//     background-size: 30px;
-//     background-position: center;
-//     background-repeat: no-repeat;
-//     border-radius: 100%;
-//     cursor: pointer;
-
-// `
